+++ CHANGELOG.md
@@ -314,3 +314,4 @@
 - 2024-03-01: Cache follow up, ready pool, discover, task, +2
 - 2024-03-02: Shutdown sender, chat, mutate, lead, +4
 - 2024-03-03: Retry follow up, lib, mod, bettercontact, +3
+- 2024-03-03: Shutdown conf, sender, descend, qualify, +3